Fix issue with smaller image on mobile device
If use c.fromDataURL(c.toDataURL()) from one canvas to another on the same page, it draws smaller image.
It can be reproduced on both, desktop browser with a scale changed, and mobile browser.

The idea of this fix is to eliminate ratio from calculations. It seems that modern browsers do it themselves.