signature_pad.js Reworked
In this completely revised version of the signature-pad for Javascript I did these changes:
First, I sorted and commented the different functions for better insight. 
Then I rewrote most of drawing logic. The original version didn't like dots, so I sorted it out following the cleanest way to do it. In Javascript, the DOWN event has the same coordinates as the first MOVE, and the last MOVE has the same as the UP event, even if there is only a single click, an MOVE event is fired. So, i add points at MOVE events, and at the UP event i decide: If the points array has four members, a bezier curve is drawn until the 3rd point, so I double the last point to draw it to the last position. Three members aren't possible, because then the first point would be doubled to start the bezier. With two members, you can only draw a single line, so I added a function for this and with only one point you need a dot - I added this also. I also consolidated some of the functions to make it cleaner.